requirements.txt
asgiref>=3.8.1
async-timeout>=5.0.1
channels>=4.2.2
channels_redis>=4.2.1
Django>=5.2.3
django-widget-tweaks>=1.5.0
dotenv>=0.9.9
msgpack>=1.1.1
pillow>=11.2.1
python-dotenv>=1.1.0
redis>=6.2.0
sqlparse>=0.5.3
typing_extensions>=4.14.0
tzdata>=2025.2
